단순화 : app.js 내 시퀄라이즈 연동코드 분리 loaders 폴더 신규 생성하여 loaders 내 정리
